% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutFLANK.R
\name{pOutlierFinderChiSqNoCorr}
\alias{pOutlierFinderChiSqNoCorr}
\title{q values for test of neutrality}
\usage{
pOutlierFinderChiSqNoCorr(DataList, Fstbar, dfInferred, qthreshold = 0.05,
  Hmin = 0.1)
}
\arguments{
\item{DataList}{A data frame with a row for each locus, that includes at least a column for $FSTNoCorr. It also helps if there is a column with an identifier for the locus. This dataframe should have empty columns called $qvalues and $OutlierFlag as well.

 @param Fstbar Mean Fst (without sample size correction) as inferred from neutral loci or OutFLank 
 
 @param dfInferred The inferred degrees of freedom of the chi-square distribution describing neutral Fst values.
 
 @param qthreshold The threshold False Discovery Rate for calling a locus an outlier ( default = 0.05)}
}
\value{
Returns a data frame with the original data, and two new columns appended:
\itemize{
\item $qvalues the q-value for a locus
\item $OutlierFlag TRUE if q is less than the qthreshold; FALSE otherwise
\item $pvalues the p-value for a locus
\item $pvaluesRightTail the one-sided (right tail) p-value for a locus
}
}
\description{
Calculates q-values for test of neutrality for a list of loci, using input of an inferred degrees of freedom for the chi-square and mean Neutral FST
}
