% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1a-data.R
\docType{data}
\name{sim1a}
\alias{sim1a}
\title{Simulated genome data}
\format{An object of class \code{"list"}}
\source{

}
\usage{
data(sim1a)
}
\description{
Data from a simulation with 6 linkage groups.
Two linkage groups are neutral, two have regions of low recombination, 
and two are neutral.
}
\examples{
data(sim1a)

\donttest{}
}
\references{
NA
}
\keyword{datasets}
