% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST functions.R
\name{WC_FST_FiniteSample_Diploids_2Alleles}
\alias{WC_FST_FiniteSample_Diploids_2Alleles}
\title{FST calculation for biallelic diploid data}
\usage{
WC_FST_FiniteSample_Diploids_2Alleles(Sample_Mat)
}
\arguments{
\item{Sample_Mat}{This is an array with a row for each population, and three values per row: Number of Homozygotes of one type, number of heterozygotes, number of homozygotes of other type.}
}
\value{
Returns a list of values related to FST:
 \itemize{
 \item   He:  the expected heterozygosity of the locus
 \item   FST:  Fst (with sample size correction)
 \item 	T1: The numerator of the Fst calculation (a from Weir and Cockerham 1984)
 \item   T2NoCorr: The denominator of the Fst calculation (a+b+c from Weir and Cockerham 1984)
 }
 @export
}
\description{
Calculates FST with correction for local sample sizes, for diploid biallelic data.
}
