% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST functions.R
\name{WC_FST_FiniteSample_Haploids_2AllelesB_MCW}
\alias{WC_FST_FiniteSample_Haploids_2AllelesB_MCW}
\title{FST calculation for biallelic haploid data}
\usage{
WC_FST_FiniteSample_Haploids_2AllelesB_MCW(AllCounts)
}
\arguments{
\item{AllCounts}{This is an array with a row for each population, and two values per row: Number of alleles in the sample of one type,  number of alleles of other type.}
}
\value{
Returns a list of values related to FST:
 \itemize{
 \item   He:  the expected heterozygosity of the locus
 \item   p_ave: the average allele frequency
 \item   FST:  Fst (with sample size correction)
 \item   T1: The numerator of the Fst calculation
 \item   T2: The denominator of the Fst calculation
 }
}
\description{
Calculates FST with correction for local sample sizes, for haploid biallelic data. Based on Weir (1996 - Genetic Data Analysis II)
}
