% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_pruned-data.R
\docType{data}
\name{which_pruned}
\alias{which_pruned}
\title{which_pruned}
\format{An object of class \code{"list"}}
\source{

}
\usage{
data(which_pruned)
}
\description{
The indexes of a quasi-independent set of SNPs for use in calculation
of neutral mean FST and df.
}
\examples{
data(which_pruned)

\donttest{}
}
\references{
NA
}
\keyword{datasets}
