% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST functions.R
\name{WC_FST_FiniteSample_Haploids_2AllelesB_NoSamplingCorrection}
\alias{WC_FST_FiniteSample_Haploids_2AllelesB_NoSamplingCorrection}
\title{FSTNoCorr calculation for biallelic haploid data}
\usage{
WC_FST_FiniteSample_Haploids_2AllelesB_NoSamplingCorrection(AllCounts)
}
\arguments{
\item{AllCounts}{This is an array with a row for each population, and two values per row: Number of alleles in the sample of one type,  number of alleles of other type.}
}
\value{
Returns a list of values related to FST:
 \itemize{
 \item   HeNoCorr:  the expected heterozygosity of the locus
 \item   p_aveNoCorr: the average allele frequency
 \item   FSTNoCorr:  Fst (without sample size correction)
 \item 	T1NoCorr: The numerator of the FSTNoCOrr calculation 
 \item   T2NoCorr: The denominator of the FSTNoCOrr calculation
 }
 @export
}
\description{
Calculates FST without correction for local sample sizes, for haploid biallelic data. This is necessary for using OutFLANK, which depends on these uncorrected values for reliable function. (Otherwise, sampling corrections can sometimes cause negative estimates of FST.)
}
